requirements.txt
python-dotenv
requests
rich
openai
typer